added cxxctp_callback_args & cxxctp_callback_result & execute_process

+++ cmake/exports/run_CXXCTP_tool.cmake
@@ -8,7 +8,9 @@ message(STATUS "output dir (for ${CXXCTP_tool_PROGRAM}) = ${OUTDIR}")
 message(STATUS "CXXCTP_tool_LOG_CONFIG (for ${CXXCTP_tool_PROGRAM}) = ${CXXCTP_tool_LOG_CONFIG}")
 
 if(CXXCTP_tool_CLEAN)
-  file(REMOVE_RECURSE ${OUTDIR})
+  if(EXISTS ${OUTDIR})
+    file(REMOVE_RECURSE ${OUTDIR})
+  endif(EXISTS ${OUTDIR})
 endif(CXXCTP_tool_CLEAN)
 
 separate_arguments(EXTRA_ARGS)
